Fixes #647. Also moved query builder api tests to sub package to catch access issues.
