+++ src/test/java/org/influxdb/querybuilder/api/AppenderTest.java
@@ -1,9 +1,11 @@
-package org.influxdb.querybuilder;
+package org.influxdb.querybuilder.api;
 
 import static org.influxdb.querybuilder.BuiltQuery.QueryBuilder.eq;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.Collections;
+
+import org.influxdb.querybuilder.Appender;
 import org.junit.jupiter.api.Test;
 
 public class AppenderTest {
